Reorganisation and add noemailer feature
