+++ back/Dockerfile
@@ -10,6 +10,10 @@ ARG DB_ROOT_PASSWORD=;
 ARG XRP_WALLET_SEED=;
 ARG XAMAN_API_KEY=;
 ARG XAMAN_SECRET_KEY=;
+ARG SMTP_HOST=;
+ARG SMTP_PORT=;
+ARG SMTP_USER=;
+ARG SMTP_PASSWORD=;
 
 WORKDIR /app
