+++ back/package.json
@@ -1,13 +1,14 @@
 {
   "dependencies": {
     "axios": "^1.6.8",
-    "ipfs-api": "^26.1.2",
-    "multer": "^1.4.5-lts.1",
     "cors": "^2.8.5",
     "dotenv": "^16.4.5",
     "express": "^4.19.2",
+    "ipfs-api": "^26.1.2",
     "jsonwebtoken": "^9.0.2",
+    "multer": "^1.4.5-lts.1",
     "mysql2": "^3.9.7",
+    "nodemailer": "^6.9.13",
     "xrpl": "^3.0.0",
     "xumm": "^1.7.7"
   },
