+++ back/paths/dbtest.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const db = require('../db');
+const db = require('../utils/db');
 
 
 // Route GET pour la racine
