+++ back/utils/mailer.js
@@ -1,0 +1,13 @@
+const nodemailer = require('nodemailer');
+
+const mailer = nodemailer.createTransport({
+    host: process.env.SMTP_HOST,
+    port: process.env.SMTP_PORT,
+    secure: false,
+    auth: {
+        user: process.env.SMTP_USER,
+        pass: process.env.SMTP_PASSWORD,
+    }
+});
+
+module.exports = mailer;
