+++ db/init.sql
@@ -6,7 +6,7 @@ CREATE TABLE
     );
 
 CREATE TABLE
-    IF NOT EXISTS NFT (
+    IF NOT EXISTS nft (
         id INT AUTO_INCREMENT PRIMARY KEY,
         userId INT NOT NULL,
         address VARCHAR(255) NOT NULL,
